pyproject.toml
[project]
name = "django-core-service"
version = "0.1.0"
description = ""
authors = [
    {name = "DenisHrtn",email = "denispython16@gmail.com"}
]
#readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (==4.2.6)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "django-celery-results (>=2.5.1,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (==4.6.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-health-check (>=3.18.3,<4.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
include = ["src/**/*"]

[tool.poetry.group.dev.dependencies]
isort = "5.12.0"
flake8 = "^7.1.1"
mypy = "^1.14.1"
pylint = "^3.3.3"
black = "^24.10.0"
pre-commit = "^4.1.0"
